+++ CHANGELOG.md
@@ -53,6 +53,8 @@
 - Add `KONG_UPSTREAM_DNS_TIME` to `kong.ctx` so that we can record the time it takes for DNS
   resolution when Kong proxies to upstream.
   [#10355](https://github.com/Kong/kong/pull/10355)
+- Support timeout for dynamic log level
+  [#10288](https://github.com/Kong/kong/pull/10288)
 
 #### Admin API
 
@@ -223,6 +225,8 @@
   [#10562](https://github.com/Kong/kong/pull/10562)
 - Bumped lua-resty-events from 0.1.3 to 0.1.4
   [#10634](https://github.com/Kong/kong/pull/10634)
+- Bumped lua-kong-nginx-module from 0.5.1 to 0.6.0
+  [#10288](https://github.com/Kong/kong/pull/10288)
 
 ## 3.2.0